# docs/bstn-api-docs/openapi.bstn-payment-api.yaml
openapi: 3.1.0
info:
  title: BSTN Payment API
  version: 1.0.0
  description: |
    Production API contract for Payment Page - BSTN Innovation Studio.
    Current stable API version: v1. The `/api/v1` response format is standardized across SNAP Link Midtrans and Dynamic QRIS.
    Versioning rule: breaking changes will use a new path version such as `/api/v2`; fields added to `/api/v1` are backwards-compatible.
servers:
  - url: https://www.bstn-innovation-studio.web.id
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/payments/{payment_id}:
    get:
      tags:
        - Payments
      summary: Get payment detail from database
      description: |
        Internal/workspace endpoint to read the local BSTN payment record.
        This endpoint does not call Midtrans and does not mutate payment status.
        Use it for audit/debugging or to retrieve the saved payment_page_url, Midtrans payment URL, and local transaction snapshot.
        For status refresh that syncs from Midtrans, use `GET /api/v1/payments/{payment_id}/status`.
      operationId: getPayment
      security:
        - ProjectBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment found in local database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ProjectBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BSTN project API key (`bstn_test_*` sandbox or `bstn_live_*` production)
  parameters:
    PaymentId:
      name: payment_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^pay_[A-Za-z0-9]+$'
      examples:
        example:
          value: pay_01HZX9T8NA2XABCDE12345
...
  responses:
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    PaymentEnvelope:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/PaymentResponse'
    PaymentResponse:
      type: object
      required:
        - payment_id
        - project_payment_ref
        - api_version
        - payment_provider
        - payment_type
        - provider_env
        - status
        - payment_mode
        - amount
        - fee_amount
        - payable_amount
        - currency
        - payment_page_url
        - created_at
        - updated_at
      properties:
        payment_id:
          type: string
          examples:
            - pay_01HZX9T8NA2XABCDE12345
        project_payment_ref:
          type: string
          examples:
            - INV-20260513-0001
        api_version:
          $ref: '#/components/schemas/ApiVersion'
        payment_provider:
          $ref: '#/components/schemas/PaymentProvider'
        payment_type:
          $ref: '#/components/schemas/PaymentType'
        provider_env:
          $ref: '#/components/schemas/ProviderEnv'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        payment_mode:
          type: string
          enum:
            - snap
            - payment_link
            - qris
        amount:
          type: integer
          examples:
            - 150000
        fee_amount:
          type: integer
          description: Difference between payable_amount and amount. Includes Dynamic QRIS unique code/admin fee or estimated SNAP customer QRIS fee for display.
        payable_amount:
          type: integer
          description: Final amount shown to customer. For Dynamic QRIS this includes unique code/admin fee; for SNAP this may include estimated customer QRIS fee while the Midtrans transaction amount remains `amount`.
        currency:
          type: string
          examples:
            - IDR
        description:
          type:
            - string
            - 'null'
        payment_page_url:
          type: string
          format: uri
        finish_url:
          type:
            - string
            - 'null'
          format: uri
        idempotent_replay:
          type: boolean
          default: false
        midtrans:
          oneOf:
            - $ref: '#/components/schemas/MidtransSummary'
            - type: 'null'
        qris:
          oneOf:
            - $ref: '#/components/schemas/QrisSummary'
            - type: 'null'
        provider:
          $ref: '#/components/schemas/ProviderSummary'
        expires_at:
          type:
            - string
            - 'null'
          format: date-time
        paid_at:
          type:
            - string
            - 'null'
          format: date-time
        canceled_at:
          type:
            - string
            - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaymentStatus:
      type: string
      enum:
        - created
        - pending
        - paid
        - challenge
        - expired
        - canceled
        - denied
        - failed
        - refunded
        - partially_refunded
        - chargeback
        - partial_chargeback
        - authorized
        - unknown
    ApiVersion:
      type: string
      enum:
        - v1
      description: API contract version used when the payment was created.
    PaymentProvider:
      type: string
      enum:
        - midtrans
        - qris_dynamic
      description: Payment gateway/provider selected for this payment.
    PaymentType:
      type: string
      enum:
        - snap_link_midtrans
        - dynamic_qr_midtrans
      description: Stable normalized payment product type.
    ProviderEnv:
      type: string
      enum:
        - sandbox
        - production
      description: Provider environment label derived from API key prefix.
    QrisSummary:
      type: object
      additionalProperties: false
      properties:
        requested:
          type: boolean
        available:
          type: boolean
        image_data_url:
          type:
            - string
            - 'null'
          description: Data URL image generated from qris_string for Dynamic QRIS when image_qris=true.
        qris_string:
          type:
            - string
            - 'null'
        dynamic_amount:
          type:
            - integer
            - 'null'
          description: Final payable amount for qris_dynamic provider after unique code and admin fee.
        payable_amount:
          type: integer
          description: Final amount customer should pay for the QR payment.
        unique_code:
          type:
            - string
            - 'null'
        admin_fee:
          type:
            - integer
            - 'null'
        admin_rate_basis_points:
          type:
            - integer
            - 'null'
        acquirer:
          type:
            - string
            - 'null'
    ProviderSummary:
      type: object
      additionalProperties: false
      properties:
        name:
          $ref: '#/components/schemas/PaymentProvider'
        environment:
          $ref: '#/components/schemas/ProviderEnv'
        payment_type:
          $ref: '#/components/schemas/PaymentType'
        transaction_id:
          type:
            - string
            - 'null'
        transaction_status:
          type:
            - string
            - 'null'
        method:
          type:
            - string
            - 'null'
        acquirer:
          type:
            - string
            - 'null'
    MidtransSummary:
      type: object
      additionalProperties: false
      properties:
        order_id:
          type:
            - string
            - 'null'
        transaction_id:
          type:
            - string
            - 'null'
        transaction_status:
          type:
            - string
            - 'null'
        payment_type:
          type:
            - string
            - 'null'
        fraud_status:
          type:
            - string
            - 'null'
        snap_token:
          type:
            - string
            - 'null'
        payment_url:
          type:
            - string
            - 'null'
          format: uri
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
    ErrorObject:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          examples:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - PAYMENT_NOT_FOUND
            - MIDTRANS_PROVIDER_ERROR
        message:
          type: string
        details:
          oneOf:
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: object
              additionalProperties: true
            - type: string
            - type: 'null'
        request_id:
          type: string
...